Fix the "closing" of tickets
history_converter uses the last modification of a ticket as a way
to "close" ticket.  In case of fake git repository generation,
closing of a ticket is removing the file.  In case of gource input
file, closing a ticket is type "|D|" log item.

Sometimes, the last change falls onto a skipped date.  Because of that,
some of the tickets may be "not closed" forever.  To avoid that, filter
the such modifications in `jira.py` instead.