history_converter.py: fix folder generation
Result of `config.sections_extensions` should be preserved as much as
possible.  The `sections = sections[:-1]` was intended for removal of
the uninteresting part of summary.

Move the removal of the last element of `sections` to avoid removal of
sections generated by user's `config.sections_extensions`.

Note that returning empty string for the empty list case is just to
avoid the unnecessary extra slash.