jira.py: simplify gathering of changes
Instead of using JIRA's JSON object of a changelog entry, gather
changelog entries as a simple tuples.

As JIRA gives timestamps for changelog entries with second precision,
storing changelogs in a list instead of mapp will have an effect of
not-squashing the entries which happened on the same ticket in the same
second.  While it probably won't make a difference in the Gource
visualization, it would produce results more "true" to the source data,
which would be easier to test and debug.